refactor(FE): `useBlockchain` hook
