+++ frontend/src/hooks/queries.js
@@ -1,30 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
-import {
-  getBlockchain,
-  LOCALSTORAGE_KEY,
-  oneTimeQueryFetchOptions,
-  QueryKeys,
-} from "../utils/utils";
-import useAuth from "./useAuth";
-import useToast from "./useToast";
-
-export function useBlockchain() {
-  const { walletConnected, setWalletAddress } = useAuth();
-  const { showToast } = useToast();
-
-  return useQuery([QueryKeys.BLOCKCHAIN], getBlockchain, {
-    enabled: walletConnected,
-    placeholderData: null,
-    onSuccess: (data) => {
-      setWalletAddress(data.signerAddress);
-      window.localStorage.setItem(LOCALSTORAGE_KEY, data.signerAddress);
-    },
-    onError: (error) => {
-      showToast(error);
-      window.localStorage.removeItem(LOCALSTORAGE_KEY);
-    },
-  });
-}
+import { oneTimeQueryFetchOptions, QueryKeys } from "../utils/utils";
+import useBlockchain from "./useBlockchain";
 
 export function useTokenName() {
   const { data: blockchain } = useBlockchain();
